+++ src/commands/settings/list.rs
@@ -8,10 +8,12 @@ use crate::model::sql::guild::*;
 async fn list(ctx: &Context, msg: &Message) -> CommandResult {
     let conf = GuildConfig::from_msg_or_respond(&ctx, msg).await?;
 
-    let s = format!("<:online:316354435745972244> Enabled\n<:offline:316354467031416832> Disabled\n\n{}", conf);
+    let s = format!(
+        "<:online:316354435745972244> Enabled\n<:offline:316354467031416832> Disabled\n\n{}",
+        conf
+    );
 
-    msg
-        .channel_id
+    msg.channel_id
         .send_message(&ctx.http, |m| {
             m.embed(|e| {
                 e.title("Guild Settings");