+++ src/handlers/roles.rs
@@ -468,7 +468,11 @@ pub async fn _message(ctx: &Context, msg: &Message) -> Result<Option<String>> {
     // get configs
     let role_config: GuildRoles = match guild_conf.role_config {
         Some(c) => serde_json::from_value(c).map_err(|e| {
-            tracing::warn!(guild_id=guild.id.0, "Failed to convert guild role config to GuildRoles struct: {}", e);
+            tracing::warn!(
+                guild_id = guild.id.0,
+                "Failed to convert guild role config to GuildRoles struct: {}",
+                e
+            );
 
             Error::Sushii("Role configuration is invalid".into())
         })?,