+++ src/model/sql/mod_log.rs
@@ -97,7 +97,7 @@ pub trait ModLogEntryDb {
         end: u64,
     ) -> Result<Vec<ModLogEntry>>;
 
-    async fn get_latest(ctx: &Context, guild_id: u64) -> Result<Option<ModLogEntry>>;
+    async fn get_latest(ctx: &Context, guild_id: u64, count: u64) -> Result<Vec<ModLogEntry>>;
 
     async fn save(&self, ctx: &Context) -> Result<ModLogEntry>;
     async fn delete(&self, ctx: &Context) -> Result<()>;
@@ -168,11 +168,11 @@ impl ModLogEntryDb for ModLogEntry {
         get_range_entries_query(pool, guild_id, start, end).await
     }
 
-    async fn get_latest(ctx: &Context, guild_id: u64) -> Result<Option<ModLogEntry>> {
+    async fn get_latest(ctx: &Context, guild_id: u64, count: u64) -> Result<Vec<ModLogEntry>> {
         let data = ctx.data.read().await;
         let pool = data.get::<DbPool>().unwrap();
 
-        get_latest_query(pool, guild_id).await
+        get_latest_query(pool, guild_id, count).await
     }
 
     /// Saves a ModLogEntry to the database. Returns a new one from the database
@@ -288,7 +288,11 @@ async fn get_range_entries_query(
     .map_err(Into::into)
 }
 
-async fn get_latest_query(pool: &sqlx::PgPool, guild_id: u64) -> Result<Option<ModLogEntry>> {
+async fn get_latest_query(
+    pool: &sqlx::PgPool,
+    guild_id: u64,
+    count: u64,
+) -> Result<Vec<ModLogEntry>> {
     sqlx::query_as!(
         ModLogEntry,
         r#"
@@ -296,11 +300,12 @@ async fn get_latest_query(pool: &sqlx::PgPool, guild_id: u64) -> Result<Option<M
                 FROM mod_logs
                WHERE guild_id = $1
             ORDER BY case_id DESC
-               LIMIT 1
+               LIMIT $2
         "#,
         guild_id as i64,
+        count as i64,
     )
-    .fetch_optional(pool)
+    .fetch_all(pool)
     .await
     .map_err(Into::into)
 }